+++ api/zalo/zalo-api.js
@@ -1,0 +1,304 @@
+// api/zalo/zalo-api.js
+import { ThreadType } from 'zca-js';
+import { zaloAccounts } from './zalo.js';
+
+export async function findUser(req, res) {
+    try {
+        const { phone, ownId } = req.body;
+        if (!phone || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        const userData = await account.api.findUser(phone);
+        res.json({ success: true, data: userData });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function getUserInfo(req, res) {
+    try {
+        const { userId, ownId } = req.body;
+        if (!userId || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        const info = await account.api.getUserInfo(userId);
+        res.json({ success: true, data: info });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function sendFriendRequest(req, res) {
+    try {
+        const { userId, ownId } = req.body;
+        if (!userId || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        const result = await account.api.sendFriendRequest('Xin chào, hãy kết bạn với tôi!', userId);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function sendMessage(req, res) {
+    try {
+        const { message, threadId, type, ownId } = req.body;
+        if (!message || !threadId || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        const msgType = type || ThreadType.User;
+        const result = await account.api.sendMessage(message, threadId, msgType);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function createGroup(req, res) {
+    try {
+        const { members, name, avatarPath, ownId } = req.body;
+        // Kiểm tra dữ liệu hợp lệ
+        if (!members || !Array.isArray(members) || members.length === 0 || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        // Gọi API createGroup từ zaloAccounts
+        const result = await account.api.createGroup({ members, name, avatarPath });
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function getGroupInfo(req, res) {
+    try {
+        const { groupId, ownId } = req.body;
+        // Kiểm tra dữ liệu: groupId phải tồn tại và nếu là mảng thì không rỗng
+        if (!groupId || (Array.isArray(groupId) && groupId.length === 0)) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        // Gọi API getGroupInfo từ zaloAccounts
+        const result = await account.api.getGroupInfo(groupId);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function addUserToGroup(req, res) {
+    try {
+        const { groupId, memberId, ownId } = req.body;
+        // Kiểm tra dữ liệu hợp lệ: groupId và memberId không được bỏ trống
+        if (!groupId || !memberId || (Array.isArray(memberId) && memberId.length === 0)) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        // Gọi API addUserToGroup từ zaloAccounts
+        const result = await account.api.addUserToGroup(memberId, groupId);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+export async function removeUserFromGroup(req, res) {
+    try {
+        const { memberId, groupId, ownId } = req.body;
+        // Kiểm tra dữ liệu: groupId và memberId phải được cung cấp, nếu memberId là mảng thì không được rỗng
+        if (!groupId || !memberId || (Array.isArray(memberId) && memberId.length === 0)) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ' });
+        }
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+        // Gọi API removeUserFromGroup từ zaloAccounts
+        const result = await account.api.removeUserFromGroup(memberId, groupId);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+// Hàm gửi một hình ảnh đến người dùng
+export async function sendImageToUser(req, res) {
+    try {
+        const { imagePath: imageUrl, threadId, ownId } = req.body;
+        if (!imageUrl || !threadId || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ: imagePath và threadId là bắt buộc' });
+        }
+
+       
+        const imagePath = await saveImage(imageUrl);
+        if (!imagePath) return res.status(500).json({ success: false, error: 'Failed to save image' });
+
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+
+        const result = await account.api.sendMessage(
+            {
+                msg: "",
+                attachments: [imagePath]
+            },
+            threadId,
+            ThreadType.User
+        ).catch(console.error);
+
+        removeImage(imagePath);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+// Hàm gửi nhiều hình ảnh đến người dùng
+export async function sendImagesToUser(req, res) {
+    try {
+        const { imagePaths: imageUrls, threadId, ownId } = req.body;
+        if (!imageUrls || !threadId || !ownId || !Array.isArray(imageUrls) || imageUrls.length === 0) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ: imagePaths phải là mảng không rỗng và threadId là bắt buộc' });
+        }
+
+      
+        const imagePaths = [];
+        for (const imageUrl of imageUrls) {
+            const imagePath = await saveImage(imageUrl);
+            if (!imagePath) {
+                // Clean up any saved images
+                for (const path of imagePaths) {
+                    removeImage(path);
+                }
+                return res.status(500).json({ success: false, error: 'Failed to save one or more images' });
+            }
+            imagePaths.push(imagePath);
+        }
+
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+
+        const result = await account.api.sendMessage(
+            {
+                msg: "",
+                attachments: imagePaths
+            },
+            threadId,
+            ThreadType.User
+        ).catch(console.error);
+
+        for (const imagePath of imagePaths) {
+            removeImage(imagePath);
+        }
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+// Hàm gửi một hình ảnh đến nhóm
+export async function sendImageToGroup(req, res) {
+    try {
+        const { imagePath: imageUrl, threadId, ownId } = req.body;
+        if (!imageUrl || !threadId || !ownId) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ: imagePath và threadId là bắt buộc' });
+        }
+
+       
+        const imagePath = await saveImage(imageUrl);
+        if (!imagePath) return res.status(500).json({ success: false, error: 'Failed to save image' });
+
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+
+        const result = await account.api.sendMessage(
+            {
+                msg: "",
+                attachments: [imagePath]
+            },
+            threadId,
+            ThreadType.Group
+        ).catch(console.error);
+
+        removeImage(imagePath);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
+
+// Hàm gửi nhiều hình ảnh đến nhóm
+export async function sendImagesToGroup(req, res) {
+    try {
+        const { imagePaths: imageUrls, threadId, ownId } = req.body;
+        if (!imageUrls || !threadId || !ownId || !Array.isArray(imageUrls) || imageUrls.length === 0) {
+            return res.status(400).json({ error: 'Dữ liệu không hợp lệ: imagePaths phải là mảng không rỗng và threadId là bắt buộc' });
+        }
+
+      
+        const imagePaths = [];
+        for (const imageUrl of imageUrls) {
+            const imagePath = await saveImage(imageUrl);
+            if (!imagePath) {
+                // Clean up any saved images
+                for (const path of imagePaths) {
+                    removeImage(path);
+                }
+                return res.status(500).json({ success: false, error: 'Failed to save one or more images' });
+            }
+            imagePaths.push(imagePath);
+        }
+
+        const account = zaloAccounts.find(acc => acc.ownId === ownId);
+        if (!account) {
+            return res.status(400).json({ error: 'Không tìm thấy tài khoản Zalo với OwnId này' });
+        }
+
+        const result = await account.api.sendMessage(
+            {
+                msg: "",
+                attachments: imagePaths
+            },
+            threadId,
+            ThreadType.Group
+        ).catch(console.error);
+
+        for (const imagePath of imagePaths) {
+            removeImage(imagePath);
+        }
+        res.json({ success: true, data: result });
+    } catch (error) {
+        res.status(500).json({ success: false, error: error.message });
+    }
+}
