+++ routes-api.js
@@ -1,0 +1,41 @@
+import express from 'express';
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { 
+    findUser, 
+    getUserInfo, 
+    sendFriendRequest, 
+    sendMessage,
+    createGroup, 
+    getGroupInfo, 
+    addUserToGroup, 
+    removeUserFromGroup,
+    sendImageToUser,
+    sendImagesToUser,
+    sendImageToGroup,
+    sendImagesToGroup
+} from './zaloService.js';
+
+const router = express.Router();
+
+// Dành cho ES Module: xác định __dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+// Giữ nguyên API cũ
+router.post('/findUser', findUser);
+router.post('/getUserInfo', getUserInfo);
+router.post('/sendFriendRequest', sendFriendRequest);
+router.post('/sendmessage', sendMessage);
+router.post('/createGroup', createGroup);
+router.post('/getGroupInfo', getGroupInfo);
+router.post('/addUserToGroup', addUserToGroup);
+router.post('/removeUserFromGroup', removeUserFromGroup);
+router.post('/sendImageToUser', sendImageToUser);
+router.post('/sendImagesToUser', sendImagesToUser);
+router.post('/sendImageToGroup', sendImageToGroup);
+router.post('/sendImagesToGroup', sendImagesToGroup);
+
+
+export default router;
